flask-int8/app/requirements.txt
Flask
requests
Pillow
numpy
opencv-python
onnx
onnxruntime
opencv-python